Refactor validator stake utilities into separate utils file
