+++ src/actions/stake/validator/index.ts
@@ -1,2 +1,3 @@
 export * from './initialize';
 export * from './stake';
+export * from './utils';
