+++ src/actions/stake/validator/initialize.ts
@@ -8,28 +8,7 @@ import {
   SystemProgram,
 } from "@solana/web3.js";
 import { STAKE_PROGRAM_ID, STAKE_CONFIG, VALIDATOR_STAKE_ACCOUNT_SIZE } from "../../../utils/constants";
-
-/**
- * Derives the validator stake PDA address using the same logic as the Rust program
- * @param validatorVote The validator vote account public key
- * @param config The stake config account public key  
- * @param programId The stake program ID
- * @returns The PDA address and bump seed
- */
-function findValidatorStakePda(
-  validatorVote: PublicKey,
-  config: PublicKey, 
-  programId: PublicKey
-): [PublicKey, number] {
-  return PublicKey.findProgramAddressSync(
-    [
-      Buffer.from("validator_stake"),
-      validatorVote.toBuffer(),
-      config.toBuffer()
-    ],
-    programId
-  );
-}
+import { findValidatorStakePda, getValidatorVoteAccount } from "./utils";
 
 /**
  * Creates an initialize instruction for validator staking
@@ -140,41 +119,4 @@ export async function makeValidatorInitializeTransaction(
   const tx = new VersionedTransaction(messageV0);
   
   return tx;
-}
-
-/**
- * Gets the vote account public key for a given validator identity
- * @param connection Solana connection instance
- * @param validatorIdentity The validator's identity public key
- * @returns The validator's vote account public key, or null if not found
- */
-export async function getValidatorVoteAccount(
-  connection: Connection,
-  validatorIdentity: PublicKey | string
-): Promise<PublicKey | null> {
-  const identityPubkey = typeof validatorIdentity === 'string' 
-    ? new PublicKey(validatorIdentity) 
-    : validatorIdentity;
-  
-  try {
-    // Get all vote accounts from the cluster
-    const voteAccounts = await connection.getVoteAccounts();
-    
-    // Search in both current and delinquent validators
-    const allValidators = [...voteAccounts.current, ...voteAccounts.delinquent];
-    
-    // Find the validator with matching identity
-    const validator = allValidators.find(
-      validator => validator.nodePubkey === identityPubkey.toBase58()
-    );
-    
-    if (validator) {
-      return new PublicKey(validator.votePubkey);
-    }
-    
-    return null;
-  } catch (error) {
-    console.error('Error fetching vote accounts:', error);
-    throw new Error(`Failed to get vote account for validator ${identityPubkey.toBase58()}: ${error}`);
-  }
 }
